Test: revamp C++ test organization
* have one executable per tested LAPACK function
* separate helper functions (e.g. matrix reassembly from solver output)
  and actual test code for future benchmarking code
