+++ TESTING/cpp/config.hpp.in
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Christoph Conrads (https://christoph-conrads.name)
+ * Copyright (c) 2020 Christoph Conrads (https://christoph-conrads.name)
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -25,9 +25,56 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
+
+#ifndef LAPACK_TESTS_CONFIG_HPP
+#define LAPACK_TESTS_CONFIG_HPP
+
 #include <boost/mpl/list.hpp>
+#include <boost/mpl/push_front.hpp>
+#include <complex>
+
+#cmakedefine01 BUILD_SINGLE
+#cmakedefine01 BUILD_DOUBLE
+#cmakedefine01 BUILD_COMPLEX
+#cmakedefine01 BUILD_COMPLEX16
 
 
-typedef boost::mpl::list<float> test_types;
+namespace lapack {
+namespace impl
+{
 
-#include <lasrtr.hpp>
+template<bool cond, class X, class List>
+struct push_if
+{
+	using type = List;
+};
+
+template<class X, class List>
+struct push_if<true, X, List>
+{
+	using type = typename boost::mpl::push_front<List, X>::type;
+};
+
+}
+
+constexpr bool BUILD_SINGLE_P = BUILD_SINGLE;
+constexpr bool BUILD_DOUBLE_P = BUILD_DOUBLE;
+constexpr bool BUILD_COMPLEX_P = BUILD_COMPLEX;
+constexpr bool BUILD_COMPLEX16_P = BUILD_COMPLEX16;
+
+using supported_types =
+	impl::push_if<BUILD_SINGLE_P, float,
+	impl::push_if<BUILD_DOUBLE_P, double,
+	impl::push_if<BUILD_COMPLEX_P, std::complex<float>,
+	impl::push_if<BUILD_COMPLEX16_P, std::complex<double>, boost::mpl::list<>
+	>::type>::type>::type>::type
+;
+
+}
+
+#undef BUILD_SINGLE
+#undef BUILD_DOUBLE
+#undef BUILD_COMPLEX
+#undef BUILD_COMPLEX16
+
+#endif
